% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{CucumberReporter}
\alias{CucumberReporter}
\title{Cucumber Reporter Base Class}
\description{
Base class for cucumber reporters that extends testthat::Reporter.
Adds step-level and feature-level reporting hooks while maintaining
full compatibility with testthat reporters.
}
\details{
CucumberReporter extends testthat's Reporter class to add:
- Feature-level hooks: `start_feature()`, `end_feature()`
- Step-level hooks: `start_step()`, `end_step()`

These additional hooks are called during cucumber test execution to provide
visibility into individual step execution. Standard testthat reporters will
continue to work unchanged, reporting only at the scenario (test) level.
}
\section{Methods}{

\describe{
  \item{`start_feature(feature_name)`}{Called before executing a feature's scenarios}
  \item{`end_feature()`}{Called after all scenarios in a feature complete}
  \item{`start_step(step)`}{Called before executing a step}
  \item{`end_step(step)`}{Called after a step completes, with status/error/duration populated}
}
}

\keyword{internal}
\section{Super class}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{CucumberReporter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{current_feature}}{Current feature name being executed}

\item{\code{current_pickle}}{Current pickle (scenario) being executed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CucumberReporter-start_feature}{\code{CucumberReporter$start_feature()}}
\item \href{#method-CucumberReporter-end_feature}{\code{CucumberReporter$end_feature()}}
\item \href{#method-CucumberReporter-start_step}{\code{CucumberReporter$start_step()}}
\item \href{#method-CucumberReporter-end_step}{\code{CucumberReporter$end_step()}}
\item \href{#method-CucumberReporter-clone}{\code{CucumberReporter$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context"><a href='../../testthat/html/Reporter.html#method-Reporter-.start_context'><code>testthat::Reporter$.start_context()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="add_result"><a href='../../testthat/html/Reporter.html#method-Reporter-add_result'><code>testthat::Reporter$add_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line"><a href='../../testthat/html/Reporter.html#method-Reporter-cat_line'><code>testthat::Reporter$cat_line()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight"><a href='../../testthat/html/Reporter.html#method-Reporter-cat_tight'><code>testthat::Reporter$cat_tight()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context"><a href='../../testthat/html/Reporter.html#method-Reporter-end_context'><code>testthat::Reporter$end_context()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context_if_started"><a href='../../testthat/html/Reporter.html#method-Reporter-end_context_if_started'><code>testthat::Reporter$end_context_if_started()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_file"><a href='../../testthat/html/Reporter.html#method-Reporter-end_file'><code>testthat::Reporter$end_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_reporter"><a href='../../testthat/html/Reporter.html#method-Reporter-end_reporter'><code>testthat::Reporter$end_reporter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_test"><a href='../../testthat/html/Reporter.html#method-Reporter-end_test'><code>testthat::Reporter$end_test()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="initialize"><a href='../../testthat/html/Reporter.html#method-Reporter-initialize'><code>testthat::Reporter$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="is_full"><a href='../../testthat/html/Reporter.html#method-Reporter-is_full'><code>testthat::Reporter$is_full()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="local_user_output"><a href='../../testthat/html/Reporter.html#method-Reporter-local_user_output'><code>testthat::Reporter$local_user_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule"><a href='../../testthat/html/Reporter.html#method-Reporter-rule'><code>testthat::Reporter$rule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_context"><a href='../../testthat/html/Reporter.html#method-Reporter-start_context'><code>testthat::Reporter$start_context()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_file"><a href='../../testthat/html/Reporter.html#method-Reporter-start_file'><code>testthat::Reporter$start_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_reporter"><a href='../../testthat/html/Reporter.html#method-Reporter-start_reporter'><code>testthat::Reporter$start_reporter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_test"><a href='../../testthat/html/Reporter.html#method-Reporter-start_test'><code>testthat::Reporter$start_test()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="update"><a href='../../testthat/html/Reporter.html#method-Reporter-update'><code>testthat::Reporter$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberReporter-start_feature"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberReporter-start_feature}{}}}
\subsection{Method \code{start_feature()}}{
Start a feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberReporter$start_feature(feature_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_name}}{Name of the feature}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberReporter-end_feature"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberReporter-end_feature}{}}}
\subsection{Method \code{end_feature()}}{
End the current feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberReporter$end_feature()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberReporter-start_step"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberReporter-start_step}{}}}
\subsection{Method \code{start_step()}}{
Start a step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberReporter$start_step(step)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{Pickle step object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberReporter-end_step"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberReporter-end_step}{}}}
\subsection{Method \code{end_step()}}{
End a step (called after execution with status/error/duration populated)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberReporter$end_step(step)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{Pickle step object with execution results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberReporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberReporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
