% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical_slice_sampler.R
\name{elliptical_slice_sample}
\alias{elliptical_slice_sample}
\title{A function that takes a positive function and returns a vector}
\usage{
elliptical_slice_sample(lx, mu = 0.2, sigma = 100, iters = 1000, start_x = 0)
}
\arguments{
\item{lx}{any function that is always positive}

\item{mu}{the mean of the normal distribution you will be sampling from}

\item{sigma}{the standard deviation of the normal distribution you will be sampling from}

\item{iters}{the number of iterations you want to sample}

\item{start_x}{the starting value of x}
}
\value{
returns a vector of x values in your sample space
}
\description{
A function that takes a positive function and returns a vector
}
