% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_samplers.R
\name{ess}
\alias{ess}
\title{Effective Sample Size}
\usage{
ess(x)
}
\arguments{
\item{x}{A numeric vector of samples from which the mean will be
estimated.}
}
\value{
A numeric scalar giving the effective sample size.
}
\description{
This function computes the effective sample size when estimating a mean based
on autocorrelated samples.
}
\examples{
ess(rnorm(100))

}
