% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence_MH.R
\name{imh_pseudo}
\alias{imh_pseudo}
\title{Independence Metropolis-Hastings Sampler}
\usage{
imh_pseudo(x, ltarget, pseudo)
}
\arguments{
\item{x}{The current state (scalar or numeric vector).}

\item{ltarget}{A function taking a scalar or numeric vector that evaluates the log-target
density, returning a numeric scalar.}

\item{pseudo}{List of length equal to the number of dimensions in \code{x}. Each element is itself a list that specifies
the pseudo-target for the corresponding dimension with functions \code{ld} that evaluates the log density
and \code{q} that evaluates the quantile (inverse-CDF) function. If the dimension of \code{x} is
one, then supply only the inner list specifying the single pseudo-target.}
}
\value{
A numeric vector with the new state.
}
\description{
Independence Metropolis-Hastings Sampler
}
\examples{
lf <- function(x) dbeta(x[1], 3, 4, log = TRUE) + dbeta(x[2], 5, 3, log = TRUE)
draws <- matrix(0.2, nrow = 10e3, ncol = 2)
nAccpt <- 0L
pseudo <- list( list(ld = function(x) dbeta(x, 2, 2, log = TRUE),
                     q = function(u) qbeta(u, 2, 2)),
                list(ld = function(x) dbeta(x, 2, 2, log = TRUE),
                     q = function(u) qbeta(u, 2, 2))
)
for (i in seq.int(2, nrow(draws))) {
 out <- imh_pseudo(draws[i - 1, ], ltarget = lf, pseudo = pseudo)
 draws[i,] <- out$x
 nAccpt <- nAccpt + out$accpt
 cat(i, '\r')
}
nAccpt / (nrow(draws) - 1)
plot(draws[,1], draws[,2], xlim = c(0, 1))
hist(draws[,1], freq = F); curve(dbeta(x, 3, 4), col = "blue", add = TRUE)
hist(draws[,2], freq = F); curve(dbeta(x, 5, 3), col = "blue", add = TRUE)
}
