% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapproxt.R
\name{lapproxt}
\alias{lapproxt}
\title{Cauchy Pseudo-Target from Laplace Approximation}
\usage{
lapproxt(f, init, sc_adj = 1, lb = -Inf, ub = Inf, ...)
}
\arguments{
\item{f}{Univariate function evaluating the unnormalized density to approximate.
Note that this is not on the log scale.}

\item{init}{Numeric scalar for an initial value (used in optimization).}

\item{sc_adj}{Positive numeric scalar; manual multiplicative adjustment to the
scale of the output pseudo-target.}

\item{lb}{Numeric scalar giving the value of left truncation of the resulting pseudo-target.
Defaults to \code{-Inf}.}

\item{ub}{Numeric scalar giving the value of right truncation of the resulting pseudo-target.
Defaults to \code{Inf}.}
}
\description{
Author: Sam Johnson
}
\examples{
pseu <- lapproxt(function(x) exp(-x^2), init = 0.5, lb = 0.0)
curve(pseu$d(x), from = -2, to = 6)
}
