% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_expectedSliceWidth.R
\name{meanSliceWidth_grid}
\alias{meanSliceWidth_grid}
\title{Expected Slice Width (over a grid)}
\usage{
meanSliceWidth_grid(x, y)
}
\arguments{
\item{x}{Numeric vector of ordered values ranging from 0 to 1.}

\item{y}{Numeric vector of evaluations of the unnormalized density at the
supplied values of \code{x}.}
}
\description{
Calculate the expected slice width of a slice sampler for an unnormalized
density supported on the unit interval.
}
\details{
Uses an approximation to the density on a grid of \code{(x, y)} pairs.
}
\examples{
meanSliceWidth_grid(x = seq(0, 1, length = 1001), y = dbeta(seq(0, 1, length = 1001), 1, 2))
}
