% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_pseudo_t.R
\name{opt_t}
\alias{opt_t}
\title{Find the Optimal Pseudo-Target for a Given Target (Student-t Family)}
\usage{
opt_t(
  target = NULL,
  samples = NULL,
  type = "samples",
  degf = c(1, 5, 20),
  lb = -Inf,
  ub = Inf,
  nbins = 100,
  coeffs = c(1, 0),
  tol_opt = 1e-06,
  tol_int = 0.001,
  plot = TRUE,
  verbose = FALSE,
  use_meanSliceWidth = FALSE
)
}
\arguments{
\item{target}{List containing functions to evaluate the unnormalized target.
Must contain function \code{ld(x)} that evaluates the log-target.}

\item{samples}{Optional numeric vector providing samples from the target distribution
(for use as alternative to \code{target}).}

\item{type}{String specifying the input type. One of "function", "samples", or "grid".
Use of "function" requires \code{target}. Use of "samples" requires \code{samples}.
Use of "grid" requires \code{target}. Defaults to "samples".}

\item{degf}{Numeric vector of degrees of freedom values to try. Defaults to \code{c(1, 5, 20)}.}

\item{lb}{Numeric scalar giving the value of left truncation. Defaults to \code{-Inf}.}

\item{ub}{Numeric scalar giving the value of right truncation. Defaults to \code{Inf}.}

\item{nbins}{Positive integer specifying the number of histogram bins if using "samples" or "grid".
Defaults to 100.}

\item{coeffs}{Positive numeric vector of length two giving relative weights of
1-the base utility (one of AUC or mean slice width) and 2-the penalty for multimodality,
measured as the area of water the density could hold. Defaults to \code{c(1.0, 0.0)}.}

\item{tol_opt}{Positive numeric scalar that passes to \code{reltol} in the call
to \code{optim()}. Defaults to \code{1.0e-6}.}

\item{tol_int}{Positive numeric scalar that passes to \code{abs.tol} in the call to \code{integrate()}.
Defaults to \code{1.0e-3}.}

\item{plot}{Logical for whether to plot a visualization of the transformed target.
Defaults to \code{TRUE}.}

\item{verbose}{Logical for whether to print intermediate steps of optimization.
Defaults to \code{FALSE}.}

\item{use_meanSliceWidth}{Logical for whether the base utility should use
Expected Slice Width (if \code{TRUE}; slower) or Area Under the Curve (AUC) if
\code{FALSE}. Defaults to \code{FALSE}.}
}
\value{
A list with named components:

\code{pseu}: a list with functions corresponding to the selected pseudo-target;
output of \code{pseudo_t_list()}.

\code{util}: a list with information about the calculated utility for the selected pseudo-target;
output of \code{utility_shrinkslice()}.

\code{opt}: output of \code{optim()}.

Other outputs repeating inputs.
}
\description{
Find the optimal pseudo-target in the Student-t family to approximate
the given (unnormalized) target. Optimize over a variety of utility functions.
Optionally supply samples from the target distribution.
}
\examples{
(pseu <- opt_t(samples = rnorm(1e3), nbins = 30, coeffs = c(1,0), plot = TRUE,
               verbose = FALSE, use_meanSliceWidth = FALSE))
(pseu <- opt_t(target = list(ld = function(x) dnorm(x, log = TRUE)),
               type = "grid", nbins = 100, coeffs = c(1,0), plot = TRUE, verbose = FALSE,
               use_meanSliceWidth = FALSE, tol_opt = 1e-3))
(pseu <- opt_t(target = list(ld = function(x) dnorm(x, log = TRUE)),
               type = "function", coeffs = c(1,0), plot = TRUE, verbose = TRUE,
               use_meanSliceWidth = FALSE, tol_opt = 1e-3, tol_int = 1e-2))
}
