% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_elliptical_slice_sampler.R
\name{single_elliptical_slice_sample}
\alias{single_elliptical_slice_sample}
\title{Slice Sampler using the Stepping Out and Shrinkage Procedures}
\usage{
single_elliptical_slice_sample(x = 0, mu = 2, sigma = 5, fx)
}
\arguments{
\item{x}{The current state (as a numeric scalar).}

\item{mu}{A numeric scalar tuning the algorithm which gives the theta value that will be used to sample a random value from the ellipse}

\item{sigma}{A numeric scalar tuning the algorithm which gives the theta value that will be used to sample a random value from the ellipse}

\item{fx}{A function taking numeric scalar and returning a numeric
scalar.}
}
\value{
A list contains two elements: "x" is the new state and "nEvaluations"
is the number of evaluations of the target function used to obtain the new
state.
}
\description{
This function implements a univariate elliptical slice sampler of Nishihara (2014)
}
