% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_sampler_generalized_ellipse}
\alias{slice_sampler_generalized_ellipse}
\title{General Elliptical Slice Sampler}
\usage{
slice_sampler_generalized_ellipse(x, target, mu, sigma, degf)
}
\arguments{
\item{x}{The current state (as a numeric scalar).}

\item{target}{A function taking numeric scalar and returning a numeric
scalar.}

\item{mu}{A numeric scalar tuning the algorithm which gives the theta value that will be used to sample a random value from the ellipse.}

\item{sigma}{A numeric scalar tuning the algorithm which gives the theta value that will be used to sample a random value from the ellipse.}

\item{degf}{Number of degrees of freedom}
}
\value{
A list contains two elements: "x" is the new state and "nEvaluations"
is the number of evaluations of the target function used to obtain the new
state.
}
\description{
General Elliptical Slice Sampler of Nishihara (2014)
}
\examples{
lf <- function(x) dbeta(x, 3, 4, log=TRUE)
draws <- numeric(1000)
nEvaluations <- 0L
system.time({
    for ( i in seq.int(2,length(draws)) ) {
        out <- slice_sampler_generalized_elliptical(draws[i-1], target=lf, mu=0.5, sigma=1, degf=5)
        draws[i] <- out$x
        nEvaluations <- nEvaluations + out$nEvaluations
    }
})
nEvaluations/length(draws)
plot(density(draws), xlim=c(0,1))
curve(exp(lf(x)), 0, 1, col="blue", add=TRUE)

}
