% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_samplers.R
\name{slice_sampler_latent}
\alias{slice_sampler_latent}
\title{Latent Slice Sampler}
\usage{
slice_sampler_latent(x, s, target, rate, log = TRUE)
}
\arguments{
\item{x}{The current state (as a numeric scalar).}

\item{s}{A random variable that determines how far the algorithm samples from
on each side}

\item{target}{A function taking numeric scalar that evaluates the target density, returning a numeric
scalar (evaluates log density if \code{log = TRUE}).}

\item{rate}{The rate parameter for a truncated exponential.}

\item{log}{If \code{FALSE}, the \code{target} function is the density
function of the target distribution.  If \code{TRUE}, the \code{target}
function evaluates the log of the density of the target distribution.}
}
\value{
A list contains two elements: "x" is the new state and "nEvaluations"
is the number of evaluations of the target function used to obtain the new
state.
}
\description{
This function implements the latent slice sampler of Li and Walker (2020).
...More details...
}
\examples{
lf <- function(x) dbeta(x, 3, 4, log = TRUE)
draws <- numeric(1000)
nEvaluations <- 0L
seconds <- system.time({
  s <- 0.5
  for (i in seq.int(2, length(draws))) {
    out <- slice_sampler_latent(draws[i - 1], s, target = lf, rate = 0.3)
    draws[i] <- out$x
    s <- out$s
    nEvaluations <- nEvaluations + out$nEvaluations
  }
})["elapsed"]
nEvaluations / length(draws)
nEvaluations / ess(draws)
ess(draws) / seconds
plot(density(draws), xlim = c(0, 1))
curve(exp(lf(x)), 0, 1, col = "blue", add = TRUE)

}
