% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_sampler_transform}
\alias{slice_sampler_transform}
\title{Transform Slice Sampler}
\usage{
slice_sampler_transform(x, log_density, pseudo_log_pdf, pseudo_inv_cdf)
}
\description{
Transform Slice Sampler
}
