% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\docType{data}
\name{step}
\alias{step}
\alias{given}
\alias{when}
\alias{then}
\title{Step}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
step

given(description, implementation)

when(description, implementation)

then(description, implementation)
}
\arguments{
\item{description}{A description of the step.}

\item{implementation}{A function that will be run when the step is executed.}
}
\description{
Step
}
\keyword{datasets}
