% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_shrinkslice_utility.R
\name{util_pseu}
\alias{util_pseu}
\title{Calculate Utility for a given Pseudo-Target}
\usage{
util_pseu(
  pseu,
  target = NULL,
  samples = NULL,
  type = "samples",
  x = NULL,
  bins = NULL,
  nbins = NULL,
  coeffs = c(1, 0),
  plot = FALSE,
  use_meanSliceWidth = TRUE,
  tol_int = 0.001
)
}
\description{
Intermediate function to bridge between \code{opt_t} and \code{utility_shrinkslice}.
}
