% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_waterArea_hist.R
\name{water_area}
\alias{water_area}
\title{Water-Holding Area Above Curve (histogram)}
\usage{
water_area(x, y, u = NULL, nbins = 30, plot = FALSE)
}
\arguments{
\item{x}{Numeric vector of histogram locations. (Not used if \code{u} is supplied).}

\item{y}{Numeric vector of histogram heights OR function evaluating the curve
for a given value of \code{u}.(Not used if \code{u} is supplied).}

\item{u}{Numeric vector of samples supported on unit interval with which to
create histogram (use \code{u = NULL} if \code{x} and \code{y} are supplied).}

\item{nbins}{Number of histogram bins to use (defaults to 30).}

\item{plot}{Logical, whether to plot a visualization of the result.}
}
\description{
Function to find how much water a curve can hold above curve (between high points)
approximated by a histogram.
This area serves as a measure of convexity/multimodality.
}
\details{
Author: Sam Johnson
}
\examples{
fn <- \(x) 0.5 * dbeta(x, 7, 2) + 0.5 * dbeta(x, 1, 6);
lf <- \(x) log(fn(x))
xx <- seq(from = 0.001, to = 0.999, length.out = 1000)
yy <- fn(xx)
water_area(x = xx, y = yy, plot = TRUE)
}
