% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_waterArea_int.R
\name{water_area_int}
\alias{water_area_int}
\title{Water-Holding Area Above Curve (function)}
\usage{
water_area_int(
  f,
  interval = c(0, 1),
  plot = FALSE,
  title = FALSE,
  eps = 0.001,
  tol_int = 0.001
)
}
\arguments{
\item{f}{Univariate function (curve) for which to evaluate water-holding area.}

\item{interval}{Numeric vector of length two bounding the domain.}

\item{plot}{Logical, whether to plot a visualization of the result.}

\item{title}{String giving a title for the plot.}

\item{eps}{Positive numeric scalar for a small buffer on the interval. Defaults to \code{1.0e-3}.}

\item{tol_int}{Positive numeric scalar that passes to \code{abs.tol} in the call to \code{integrate()}.
Defaults to \code{1.0e-3}.}
}
\value{
List with:

\code{totalArea}: Total area of a rectangle enclosing the curve (width is the
length of the interval, height is the max of \code{f} in the interval)

\code{AUC}: Area under the curve (not restricted to unit square)

\code{totalWaterArea}: Total water-holding area (not restricted to unit square)
}
\description{
Function to find how much water a curve can hold above curve (between high points).
This area serves as a measure of convexity/multimodality.
}
\details{
Author: Sam Johnson
}
\examples{
f <- function(x) {0.33 * dnorm(x,0,0.5) + 0.33 * dnorm(x,2,1) + 0.33 * dnorm(x,7,2)}
interval <- c(-3,15)
water_area_int(f = f, interval = interval, plot = TRUE)
}
