% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{CucumberProgressReporter}
\alias{CucumberProgressReporter}
\title{Progress Reporter for Cucumber}
\description{
A reporter that prints step-by-step progress to the console, mimicking
testthat::ProgressReporter but with visibility into individual steps.
}
\details{
Prints each step as it executes with a status indicator:
- ✓ for passed steps (green)
- ✗ for failed/errored steps (red)

Groups steps under scenario and feature headers, similar to testthat's output.
}
\examples{
\dontrun{
# Use with cucumber::test()
cucumber::test(
  "tests/acceptance",
  reporter = cucumber::CucumberProgressReporter$new()
)

# Or with verbose output
cucumber::test(
  "tests/acceptance",
  reporter = cucumber::CucumberProgressReporter$new(show_praise = FALSE)
)
}

}
\section{Super classes}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{\link[cucumber:CucumberReporter]{cucumber::CucumberReporter}} -> \code{CucumberProgressReporter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{show_praise}}{Whether to show praise messages}

\item{\code{step_count}}{Number of steps executed}

\item{\code{step_passed}}{Number of steps that passed}

\item{\code{step_failed}}{Number of steps that failed}

\item{\code{current_scenario}}{Name of the scenario currently executing}

\item{\code{current_steps}}{Steps executed so far in the current scenario}

\item{\code{failures}}{Recorded failures for the end-of-run summary}

\item{\code{step_errored}}{Whether end_step already rendered an error this scenario}

\item{\code{num_colors}}{Terminal color support captured at construction}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CucumberProgressReporter-new}{\code{CucumberProgressReporter$new()}}
\item \href{#method-CucumberProgressReporter-start_feature}{\code{CucumberProgressReporter$start_feature()}}
\item \href{#method-CucumberProgressReporter-end_feature}{\code{CucumberProgressReporter$end_feature()}}
\item \href{#method-CucumberProgressReporter-start_test}{\code{CucumberProgressReporter$start_test()}}
\item \href{#method-CucumberProgressReporter-add_result}{\code{CucumberProgressReporter$add_result()}}
\item \href{#method-CucumberProgressReporter-start_step}{\code{CucumberProgressReporter$start_step()}}
\item \href{#method-CucumberProgressReporter-end_step}{\code{CucumberProgressReporter$end_step()}}
\item \href{#method-CucumberProgressReporter-end_reporter}{\code{CucumberProgressReporter$end_reporter()}}
\item \href{#method-CucumberProgressReporter-clone}{\code{CucumberProgressReporter$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context"><a href='../../testthat/html/Reporter.html#method-Reporter-.start_context'><code>testthat::Reporter$.start_context()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line"><a href='../../testthat/html/Reporter.html#method-Reporter-cat_line'><code>testthat::Reporter$cat_line()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight"><a href='../../testthat/html/Reporter.html#method-Reporter-cat_tight'><code>testthat::Reporter$cat_tight()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context"><a href='../../testthat/html/Reporter.html#method-Reporter-end_context'><code>testthat::Reporter$end_context()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context_if_started"><a href='../../testthat/html/Reporter.html#method-Reporter-end_context_if_started'><code>testthat::Reporter$end_context_if_started()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_file"><a href='../../testthat/html/Reporter.html#method-Reporter-end_file'><code>testthat::Reporter$end_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_test"><a href='../../testthat/html/Reporter.html#method-Reporter-end_test'><code>testthat::Reporter$end_test()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="is_full"><a href='../../testthat/html/Reporter.html#method-Reporter-is_full'><code>testthat::Reporter$is_full()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="local_user_output"><a href='../../testthat/html/Reporter.html#method-Reporter-local_user_output'><code>testthat::Reporter$local_user_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule"><a href='../../testthat/html/Reporter.html#method-Reporter-rule'><code>testthat::Reporter$rule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_context"><a href='../../testthat/html/Reporter.html#method-Reporter-start_context'><code>testthat::Reporter$start_context()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_file"><a href='../../testthat/html/Reporter.html#method-Reporter-start_file'><code>testthat::Reporter$start_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_reporter"><a href='../../testthat/html/Reporter.html#method-Reporter-start_reporter'><code>testthat::Reporter$start_reporter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="update"><a href='../../testthat/html/Reporter.html#method-Reporter-update'><code>testthat::Reporter$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberProgressReporter-new"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberProgressReporter-new}{}}}
\subsection{Method \code{new()}}{
Initialize the reporter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberProgressReporter$new(show_praise = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{show_praise}}{Whether to show praise (default TRUE)}

\item{\code{...}}{Additional arguments passed to parent}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberProgressReporter-start_feature"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberProgressReporter-start_feature}{}}}
\subsection{Method \code{start_feature()}}{
Start a feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberProgressReporter$start_feature(feature_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_name}}{Name of the feature}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberProgressReporter-end_feature"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberProgressReporter-end_feature}{}}}
\subsection{Method \code{end_feature()}}{
End the current feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberProgressReporter$end_feature()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberProgressReporter-start_test"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberProgressReporter-start_test}{}}}
\subsection{Method \code{start_test()}}{
Start a test (called by testthat for each scenario)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberProgressReporter$start_test(context, test)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context}}{Test context}

\item{\code{test}}{Test name (will be "Scenario: <name>")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberProgressReporter-add_result"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberProgressReporter-add_result}{}}}
\subsection{Method \code{add_result()}}{
Add a result (warning, skip, failure, etc.)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberProgressReporter$add_result(context, test, result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context}}{Test context}

\item{\code{test}}{Test object}

\item{\code{result}}{Test result object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberProgressReporter-start_step"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberProgressReporter-start_step}{}}}
\subsection{Method \code{start_step()}}{
Start a step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberProgressReporter$start_step(step)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{Pickle step object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberProgressReporter-end_step"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberProgressReporter-end_step}{}}}
\subsection{Method \code{end_step()}}{
End a step and print its result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberProgressReporter$end_step(step)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{Pickle step object with execution results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberProgressReporter-end_reporter"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberProgressReporter-end_reporter}{}}}
\subsection{Method \code{end_reporter()}}{
Print summary at the end (called by testthat)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberProgressReporter$end_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CucumberProgressReporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CucumberProgressReporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CucumberProgressReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
