% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-opts.R
\name{opts}
\alias{opts}
\title{{cucumber} Options}
\description{
Internally used, package-specific options.
They allow overriding the default behavior of the package.
}
\details{
The following options are available:
\itemize{
\item \code{cucumber.indent}

Regular expression for the indent of the feature files.

default: \verb{^\\\\s\{2\}}
\item \code{cucumber.interactive}

Logical value indicating whether to ask which feature files to run.

default: \code{FALSE}
}

See \code{\link[base:options]{base::options()}} and \code{\link[base:options]{base::getOption()}} on how to work with options.
}
