% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step}
\alias{step}
\alias{given}
\alias{when}
\alias{then}
\title{Step}
\usage{
given(description, implementation)

when(description, implementation)

then(description, implementation)
}
\arguments{
\item{description}{A description of the step.}

\item{implementation}{A function that will be run when the step is executed.}
}
\value{
None, funcion called for side effects.
}
\description{
The implementation function should always have the last parameter named `context`.
It holds the environment where state is stored.
}
\details{
Named arguments that precede `context` are considered parameters.

If a step has a description "I have {int} cucumbers in my basket" then the implementation
function should be a `function(n_cucumbers, context)`. The {int} value will be passed to
`n_cucumbers`, this parameter can have any name.

If a step has a table, then the table will be passed as an argument next after inline parameters
and before `context`. See `inst/examples/table` how write implemntation that uses tables.
}
