% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Run all Cucumber tests}
\usage{
test(
  features_dir = getOption("cucumber_features_dir", "tests/testthat"),
  steps_dir = getOption("cucumber_steps_dir", "tests/testthat/steps"),
  steps_loader = function(steps_dir) {
     steps_files <- dir_ls(steps_dir, glob =
    "*.R$", type = "file")
     walk(steps_files, source)
     get_steps()
 }
)
}
\arguments{
\item{features_dir}{A character string of the directory containing the feature files.
By default it checks "tests/testthat" directory. To setup a different, default directory
use `option(cucumber_features_dir = "path/to/your/features")`.}

\item{steps_dir}{A character string of the directory containing the step files.
By default it checks "tests/testthat/steps" directory. To setup a different, default directory
use `option(cucumber_steps_dir = "path/to/your/steps")`.}

\item{steps_loader}{A function that loads the steps implementations. By default it sources all files
from the \code{steps_dir} using the built-in mechanism. You can provide your own function to load
the steps. The function should take one argument, which will be the \code{steps_dir}
and return a list of steps.}
}
\description{
This command runs all Cucumber tests. It takes all .feature files
from the `features_dir` and runs them using the steps from the `steps_dir`.
}
