% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test}
\usage{
test(features_dir = "tests/testthat", steps_dir = "tests/testthat/steps")
}
\arguments{
\item{features_dir}{A character string of the directory containing the feature files.}

\item{steps_dir}{A character string of the directory containing the step files.}
}
\description{
Test
}
