% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Run all Cucumber tests}
\usage{
test(features_dir, steps_dir, steps_loader = .default_steps_loader)
}
\arguments{
\item{features_dir}{A character string of the directory containing the feature files.}

\item{steps_dir}{A character string of the directory containing the step files.}

\item{steps_loader}{A function that loads the steps implementations. By default it sources all files
from the \code{steps_dir} using the built-in mechanism. You can provide your own function to load
the steps. The function should take one argument, which will be the \code{steps_dir}
and return a list of steps.}
}
\value{
None, function called for side effects.
}
\description{
This command runs all Cucumber tests. It takes all .feature files
from the \code{features_dir} and runs them using the steps from the \code{steps_dir}.
}
